+++ src/main.rs
@@ -7,7 +7,7 @@ use assassin::broker::Broker;
 use assassin::commission::charles_schwab::CharlesSchwab;
 
 static INPUT_FILE: &'static str = "/Users/billrobinson/Desktop/aapl_2013.csv";
-static STARTING_CAPITAL: f64 = 100_000.0;
+static STARTING_CAPITAL: f32 = 100_000.0;
 
 fn main() {
 	let feed = DiscountOptionData::new(INPUT_FILE);
